rename upgrade handler
