+++ app/upgrades/v14rc5/upgrades.go
@@ -1,7 +1,7 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
 
-package v14rc4
+package v14rc5
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
